fix: double-precision floating point used for monetary amounts causing rounding errors and drift in customer outstanding balances and revenue records

+++ src/main/java/com/example/waterrefillapijava/dto/DeliveryRequest.java
@@ -1,5 +1,7 @@
 package com.example.waterrefillapijava.dto;
 
+import java.math.BigDecimal;
+
 import com.example.waterrefillapijava.model.DeliveryStatus;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
@@ -14,6 +16,6 @@ public record DeliveryRequest(
 	Integer bottlesReturned,
 
 	@JsonProperty("cash_collected")
-	Double cashCollected
+	BigDecimal cashCollected
 ) {
 }
